Refactor/restructure examples

+++ pynektools/interpolation/mesh_to_mesh.py
@@ -150,9 +150,9 @@ class PMapper:
         self.nelv = msh.nelv
 
         # Allocate the new coordinates
-        x = np.zeros((msh.nelv, self.lz, self.ly, self.lx))
-        y = np.zeros((msh.nelv, self.lz, self.ly, self.lx))
-        z = np.zeros((msh.nelv, self.lz, self.ly, self.lx))
+        x = np.zeros((msh.nelv, self.lz, self.ly, self.lx), dtype=msh.x.dtype)
+        y = np.zeros((msh.nelv, self.lz, self.ly, self.lx), dtype=msh.x.dtype)
+        z = np.zeros((msh.nelv, self.lz, self.ly, self.lx), dtype=msh.x.dtype)
 
         # Loop over the elements and perform the interpolation
         x_gll = self.r_dist
@@ -183,7 +183,7 @@ class PMapper:
         # Allocate the result of the interpolation
         interpolated_fields = []
         for _ in range(0, number_of_fields):
-            interpolated_fields.append(np.zeros((self.nelv, self.lz, self.ly, self.lx)))
+            interpolated_fields.append(np.zeros((self.nelv, self.lz, self.ly, self.lx), dtype=field_list[0].dtype))
 
         # Get the RST coordinates of the new points
         x_gll = self.r_dist
